pygui/requirements.txt
PyQt5PyDub
Pillow